Fixed History button
created history button in MenuBarView and it that way that it will be added to prog content panel through new method I made in menubar view. thus the menubar controller is the one that will handle history button action.
I suspect person who will do history page will not need to use that button, but in case, let me know.